move checksum to int variable

+++ src/main/java/games/rednblack/h2d/extension/spine/SlotBody.java
@@ -3,6 +3,6 @@ package games.rednblack.h2d.extension.spine;
 import com.badlogic.gdx.physics.box2d.Body;
 
 public class SlotBody {
-    public float checksum;
+    public int checksum;
     public Body body;
 }
